(Dub) Updated Airing Schedule
Something is wrong! There are 154 dub titles resolved and there are 155 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 8 for A Star Brighter Than the Sun
(Dub) Added Episode 8 for A Star Brighter Than the Sun
(Dub) Added Episode 8 for A Star Brighter Than the Sun
(Dub) Added Episode 8 for A Star Brighter Than the Sun