(Dub) Updated Airing Schedule
Something is wrong! There are 154 dub titles resolved and there are 155 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 19 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 19 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 19 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 19 for Watari-kun's ****** Is about to Collapse