(Dub) Updated Airing Schedule
Something is wrong! There are 149 dub titles resolved and there are 151 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 6 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 6 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 6 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 6 for The Banished Court Magician Aims to Become the Strongest