(Dub) Updated Airing Schedule
Something is wrong! There are 149 dub titles resolved and there are 151 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 9 for A Star Brighter Than the Sun
(Dub) Added Episode 9 for A Star Brighter Than the Sun
(Dub) Added Episode 9 for A Star Brighter Than the Sun
(Dub) Added Episode 9 for A Star Brighter Than the Sun