(Dub) Updated Airing Schedule
Something is wrong! There are 149 dub titles resolved and there are 151 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 20 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 20 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 20 for Watari-kun's ****** Is about to Collapse
(Dub) Added Episode 20 for Watari-kun's ****** Is about to Collapse