(Dub) Updated Airing Schedule
Something is wrong! There are 139 dub titles resolved and there are 142 dub titles in the timetables, less than what is expected!
(Dub) Added Episode 9 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 9 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 9 for The Banished Court Magician Aims to Become the Strongest
(Dub) Added Episode 9 for The Banished Court Magician Aims to Become the Strongest