(Dub) Updated Airing Schedule
Something is wrong! There are 105 dub titles resolved and there are 109 dub titles in the timetables, less than what is expected!
(Dub) Removed Episode 11 of Kakuriyo -Bed & Breakfast for Spirits- 2 as it has been delayed
(Dub) Removed Episode 11 of Kakuriyo -Bed & Breakfast for Spirits- 2 as it has been delayed
(Dub) Removed Episode 11 of Let’s Play as it has been delayed
(Dub) Removed Episode 11 of Let’s Play as it has been delayed
(Dub) Removed Episode 10 of The Dark History of the Reincarnated Villainess as it has been delayed
(Dub) Removed Episode 10 of The Dark History of the Reincarnated Villainess as it has been delayed
(Dub) Removed Episode 10 of The Dark History of the Reincarnated Villainess as it has been delayed
(Dub) Removed Episode 4 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 4 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 4 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 3 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 3 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 3 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 2 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 2 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 2 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 1 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 1 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Removed Episode 1 of DIGIMON BEATBREAK due to a correction in the airing date
(Dub) Added Episode 1 for DIGIMON BEATBREAK
(Dub) Added Episode 1 for DIGIMON BEATBREAK
(Dub) Added Episode 1 for DIGIMON BEATBREAK
(Dub) Added Episode 2 for DIGIMON BEATBREAK
(Dub) Added Episode 2 for DIGIMON BEATBREAK
(Dub) Added Episode 2 for DIGIMON BEATBREAK
(Dub) Added Episode 3 for DIGIMON BEATBREAK
(Dub) Added Episode 3 for DIGIMON BEATBREAK
(Dub) Added Episode 3 for DIGIMON BEATBREAK
(Dub) Added Episode 4 for DIGIMON BEATBREAK
(Dub) Added Episode 4 for DIGIMON BEATBREAK
(Dub) Added Episode 4 for DIGIMON BEATBREAK